fix(types): carry 64-bit integers as exact text
JavaScript numbers are doubles, exact only to 2^53, so every Firebird
BIGINT lost its tail on the way to the UI. Verified: the maximum
BIGINT 9223372036854775807 arrived as 9223372036854776000, and
9007199254740993 came back one lower. This hit BIGINT primary keys,
generators and MON$ counters — ordinary columns, silently wrong.

Adds `plamenix_types::exact_int`, a serde adapter writing an i64 as
decimal text, and applies it to the two integers whose magnitude the
user controls: `ColumnValue::Integer` and
`GeneratorInfo.current_value`. Both are `#[specta(type = String)]` so
the generated TypeScript says `string`. Deserialisation still accepts
a JSON number, so payloads written before this change load.

The ts-gen remap that turned i64 into `number` stays for the counters
that are bounded by construction, but its comment claimed every
Plamenix integer fits in 2^53, which was the false premise behind the
bug. Rewritten to say which integers it may cover and where the
exact ones are declared instead.

+++ crates/plamenix-db/src/query.rs
@@ -32,7 +32,15 @@ pub enum ColumnValue {
     /// Textual data (CHAR, VARCHAR, BLOB `SUB_TYPE` TEXT).
     Text(String),
     /// 64-bit signed integer (SMALLINT, INTEGER, BIGINT).
-    Integer(i64),
+    ///
+    /// Carried as decimal text: the magnitude is the user's data, and a
+    /// `BIGINT` past 2^53 cannot survive a JSON number. See
+    /// [`plamenix_types::exact_int`].
+    Integer(
+        #[serde(with = "plamenix_types::exact_int")]
+        #[specta(type = String)]
+        i64,
+    ),
     /// Double-precision floating point (FLOAT, DOUBLE PRECISION).
     Float(f64),
     /// Boolean (FB 3.0+ `BOOLEAN`).
